moved all form interfaces to the separate folder
