+++ src/interfaces/searchForm/SearchFormState.tsx
